Revert "Implementando botão de salvar em cadastro de categorias."
This reverts commit 596b39b7146298687929a6c0c3663eba1cdb26f2.
