+++ src/pages/acessibility-guide/AcessibilityGuide.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { AxiosResponse } from 'axios';
 import { act, render, screen } from '@testing-library/react';
-import CardGuidesResponse, { getGuides } from '@services/guides';
+import  CardGuidesResponse, { getGuides } from '@services/guides';
 import AcessibilityGuide from '@pages/acessibility-guide';
 import '@testing-library/jest-dom';
