+++ src/pages/register-digital-content/index.tsx
@@ -3,8 +3,7 @@ import React from 'react';
 export interface RegisterDigitalContentProps {}
 
 export const RegisterDigitalContent: React.FC<
-  RegisterDigitalContentProps
-> = (): JSX.Element => {
+  RegisterDigitalContentProps> = (): JSX.Element => {
   return <p>Teste de página de Cadastro de Conteúdo Digital</p>;
 };
