+++ src/pages/register-guide/RegisterGuide.test.tsx
@@ -66,7 +66,7 @@ describe('Página de cadastro de nova guia', () => {
     // eslint-disable-next-line testing-library/no-unnecessary-act
     act(() => {
       render(<RegisterGuide />);
-    });
+      });
 
     const textoBotaoSubmit = 'Salvar';
     const botaoSubmit = screen.getByText(textoBotaoSubmit);
@@ -82,7 +82,7 @@ describe('Página de cadastro de nova guia', () => {
   test('Deve chamar a função postGuides quando o botão do submit for clicado', () => {
     // eslint-disable-next-line testing-library/no-unnecessary-act
     act(() => {
-      render(<RegisterGuide />);
+    render(<RegisterGuide />);
     });
     const textoNoBotaoSubmit = 'Salvar';
     const botaoSubmit = screen.getByText(textoNoBotaoSubmit);
@@ -97,7 +97,7 @@ describe('Página de cadastro de nova guia', () => {
     // eslint-disable-next-line testing-library/no-unnecessary-act
     act(() => {
       render(<RegisterGuide />);
-    });
+      });
 
     validateInputMock.mockResolvedValue(true as unknown as InputInterface);
     postGuidesMock.mockResolvedValue(true as unknown as Promise<AxiosResponse>);
@@ -117,10 +117,10 @@ describe('Página de cadastro de nova guia', () => {
   test('Deve mostrar na tela o card de notificação de erro quando o botão de submit for clicado', async () => {
     act(() => {
       render(<RegisterGuide />);
-    });
+      });
 
-    const errorMessage = 'Erro';
-    const throwError = new Error(errorMessage);
+      const errorMessage = "Erro";
+      const throwError = new Error(errorMessage);
 
     validateInputMock.mockImplementation(() => {
       throw throwError;
@@ -129,7 +129,7 @@ describe('Página de cadastro de nova guia', () => {
     const textoNoBotaoSubmit = 'Salvar';
     const NotificationMessage = errorMessage;
     const botaoSubmit = screen.getByText(textoNoBotaoSubmit);
-
+    
     act(() => {
       userEvent.click(botaoSubmit);
     });
