+++ src/pages/register-guide/index.tsx
@@ -31,9 +31,9 @@ export const RegisterGuide: React.FC<RegisterGuideProps> = (): JSX.Element => {
       setErrorMessage(error.message);
       setError(true);
     }
-
-    title.current!.value = '';
-    description.current!.value = '';
+    
+    title.current!.value = "";
+    description.current!.value = "";
   }
 
   return (
